decimals

+++ DataLayer/Models2/ZraStockMaster.cs
@@ -16,12 +16,12 @@ namespace DataLayer.Models2
         public decimal Quantity { get; set; }
         public string? TaxLabel { get; set; }
         public string? BranchId { get; set; }
-        public decimal? Prc { get; set; } = 0;
-        public decimal? SplyAmt { get; set; } = 0;
+        public double? Prc { get; set; } = 0;
+        public double? SplyAmt { get; set; } = 0;
         public decimal? TaxblAmt { get; set; } = 0;
         public string? VatCatCd { get; set; } = "";
         public decimal? TaxAmt { get; set; } = 0;
-        public decimal? TotAmt { get; set; } = 0;
+        public double? TotAmt { get; set; } = 0;
     }
 }
